+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changelog
  * renamed `InstallationParams::getPublicPath()` to `getServerPath()`
  * upgraded to webmozart/path-util 2.0
  * fixed `PackageFileInstallerManager` for unloadable packages
+ * decoupled from puli/factory
 
 * 1.0.0-beta4 (2015-04-13)